+++ src/components/features/draft/DraftPage/components/actions/FloatingActionButton/index.stories.tsx
