+++ src/components/features/draft/DraftPage/components/chat/ChatLogSection/LogItem/index.tsx
