+++ src/components/features/draft/DraftPage/components/chat/index.tsx
@@ -1,0 +1,1 @@
+export { ChatLogSection } from './ChatLogSection';
